+++ src/const.js
@@ -76,7 +76,17 @@ module.exports = {
     ERROR: 1,
     NONE: 0,
   },
+
+  /**
+   * Deletion Modes
+   * @property {Object} [DELETION_MODE=]
+   * @property {number} DELETION_MODE.ALL          Delete Message/Conversation for All users but remain in the Conversation;
+   *                                               new Messages will restore this Conversation minus any Message History prior to deletion.
+   * @property {number} DELETION_MODE.MY_DEVICES   Delete Message or Conversation; but see layer.Conversation.leave if you want to delete
+   *                                               a Conversation and not have it come back.
+   */
   DELETION_MODE: {
     ALL: 1,
+    MY_DEVICES: 2,
   },
 };